+++ lib/console_kit/connections/base_connection_handler.rb
@@ -14,7 +14,13 @@ module ConsoleKit
 
       def initialize(context) = @context = context
       def connect = raise NotImplementedError, "#{self.class} must implement #connect"
-      def available? = false
+      def available? = raise NotImplementedError, "#{self.class} must implement #available?"
+
+      private
+
+      def context_attribute(name)
+        @context.respond_to?(name, true) ? @context.send(name) : nil
+      end
     end
   end
 end