+++ lib/console_kit/connections/connection_manager.rb
@@ -2,6 +2,8 @@
 
 require_relative 'sql_connection_handler'
 require_relative 'mongo_connection_handler'
+require_relative 'redis_connection_handler'
+require_relative 'elasticsearch_connection_handler'
 
 module ConsoleKit
   module Connections
@@ -12,6 +14,8 @@ module ConsoleKit
           handler_classes.filter_map do |klass|
             handler = klass.new(context)
             handler if handler.available?
+          rescue NotImplementedError
+            nil
           end
         end