+++ lib/console_kit/connections/mongo_connection_handler.rb
@@ -1,22 +1,17 @@
 # frozen_string_literal: true
 
-require 'forwardable'
 require_relative 'base_connection_handler'
 
 module ConsoleKit
   module Connections
     # Handles MongoDB connections
     class MongoConnectionHandler < BaseConnectionHandler
-      extend Forwardable
-
-      def_delegator :@context, :tenant_mongo_db
-
       def connect
-        db = tenant_mongo_db.presence
+        db = context_attribute(:tenant_mongo_db).presence
         Output.print_info(switch_message(db))
-        Mongoid.override_client(db)
+        Mongoid.override_database(db)
       rescue NoMethodError
-        Output.print_warning('Mongoid.override_client is not defined.')
+        Output.print_warning('Mongoid.override_database is not available in this version of Mongoid.')
       end
 
       def available? = defined?(Mongoid)